pyproject.toml
[tool]
[tool.poetry]
name = "setu"
version = "1.0.0"
homepage = "https://github.com/SetuHQ/setu-python-sdk"
description = "Python package to connect to Setu's UPI Deep Link APIs."
authors = ["Naresh R <ghostwriternr@gmail.com>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "setu" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
PyJWT = "^2.4.0"
requests = "^2.27.1"
marshmallow = "^3.14.1"
marshmallow-oneofschema = "^3.0.1"

[tool.poetry.dev-dependencies]
black  = "22.3.0"
isort = "^5.8.0"
flake8 = "^3.9.2"
flake8-docstrings = "^1.6.0"
mypy = "^0.950"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
tox = "^3.20.1"
tox-pyenv = "^1.1.0"
virtualenv = "^20.2.2"
pip = "^20.3.1"
mkdocs = "^1.3.0"
mkdocs-include-markdown-plugin = "^3.4.0"
mkdocs-material = "^8.2.7"
mkdocstrings = "^0.17.0"
mkdocs-material-extensions = "^1.0.1"
twine = "^3.8.0"
mkdocs-autorefs = "^0.3.1"
pre-commit = "^2.17.0"
toml = "^0.10.2"
bump2version = "^1.0.1"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
